+++ src/ext/coms_curl.c
@@ -276,6 +276,7 @@ static void *writer_loop(void *_) {
 
     _curl_free(writer);
     _signal_writer_finished(writer);
+    ddtrace_coms_shutdown();
     return NULL;
 }
 
@@ -402,10 +403,11 @@ BOOL_T ddtrace_coms_on_request_finished() {
     return TRUE;
 }
 
+// Returns TRUE if writer is shutdown completely
 BOOL_T ddtrace_coms_flush_shutdown_writer_synchronous() {
     struct _writer_loop_data_t *writer = get_writer();
     if (!writer->thread) {
-        return FALSE;
+        return TRUE;
     }
 
     writer_set_shutdown_state(writer);
@@ -435,8 +437,9 @@ BOOL_T ddtrace_coms_flush_shutdown_writer_synchronous() {
         pthread_join(writer->thread->self, NULL);
         free(writer->thread);
         writer->thread = NULL;
+        return TRUE;
     }
-    return TRUE;
+    return FALSE;
 }
 
 BOOL_T ddtrace_coms_synchronous_flush(uint32_t timeout) {